added CSV FileSource Connector and removed unnecessary configurations and containers
