+++ influxdb/create_influxdb.sh
@@ -7,7 +7,7 @@ set -e
 
 # Send POST request to create schema
 #echo 'Waiting for server start...'
-#sleep 45
+sleep 45
 echo 'Creating influxdb-sink connector...'
 
-docker-compose run kafkaconnect create influxdb-sink -d mydb foo
+docker-compose run kafkaconnect create influxdb-sink -d smartwatch_db db_upload
